+++ Demo/Shared/Debug/ViewDebug.swift
@@ -6,10 +6,29 @@
 //  Copyright © 2022 Daniel Saidi. All rights reserved.
 //
 
-import Foundation
+import SwiftUI
 
 final class ViewDebug {
 
-    /// Set this to true to enable view debugging.
+    /**
+     Make any view draw a random background color every time
+     it's redrawn.
+     */
     static var isEnabled = false
+}
+
+public extension View {
+
+    /**
+     Make the view draw a random background color every time
+     it's redrawn, but only if `ViewDebug.isEnabled` is true.
+     */
+    @ViewBuilder
+    func viewDebug() -> some View {
+        if ViewDebug.isEnabled {
+            self.background(Color.random())
+        } else {
+            self
+        }
+    }
 }