+++ Sources/RichTextKit/_Deprecated/RichTextDataWriter+Deprecated.swift
@@ -1,5 +1,12 @@
 import Foundation
 
+@available(*, deprecated, message: "Use RichTextDataReader instead")
+public protocol RichTextDataWriter: RichTextDataReader {}
+
+@available(*, deprecated, message: "Use RichTextDataReader instead")
+extension NSAttributedString: RichTextDataWriter {}
+
+@available(*, deprecated, message: "Use RichTextDataReader instead")
 public extension RichTextDataWriter {
 
     @available(*, deprecated, renamed: "richTextData(for:)")